Major performance overhaul - Refactored cells to contain all the information for lighting and visibility - Added copy constructor and static const null glyph (no ascii character and black colours) - Refactored rendering to improve performance - Added timer for drawing to limit FPS to 60 - Refactored camera logic for improved performance - Added visual studio profiling files to .gitignore

+++ include/map.h
@@ -16,7 +16,7 @@ public:
 
 	virtual ~Map(void) = default;
 
-	int size(void)   const { return _items.size(); }
+	size_t size(void)   const { return _items.size(); }
 	int width(void)  const { return _width; }
 	int height(void) const { return _height; }
